# packages/sync-protocol/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Platform API
  version: 0.1.0
  description: |
    Multi-tenant inventory ledger + sync. **Card tenders require an active connection to the central API**
    by default; offline checkout is allowed for cash (and other offline-capable tenders) per tenant policy.
servers:
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /v1/admin/auth/login:
    post:
      tags: [Admin]
      summary: Operator login (JWT for admin routes)
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Bearer access token
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_in]
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
...
  /v1/admin/shops/{shop_id}/default-tax:
    patch:
      tags: [Admin]
      summary: Set shop default tax rate (basis points)
      operationId: adminPatchShopDefaultTax
      security:
        - bearerAuth: []
      parameters:
        - name: shop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [default_tax_rate_bps]
              properties:
                default_tax_rate_bps:
                  type: integer
                  minimum: 0
                  maximum: 100000
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  shop_id:
                    type: string
                    format: uuid
                  default_tax_rate_bps:
                    type: integer
  /v1/admin/shops/{shop_id}/product-taxes:
    get:
      tags: [Admin]
      summary: List per-product tax overrides for a shop
      operationId: adminListShopProductTaxes
      security:
        - bearerAuth: []
      parameters:
        - name: shop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShopProductTaxRecord"
  /v1/admin/shops/{shop_id}/product-taxes/{product_id}:
    put:
      tags: [Admin]
      summary: Upsert per-product tax override
      operationId: adminUpsertShopProductTax
      security:
        - bearerAuth: []
      parameters:
        - name: shop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tax_exempt]
              properties:
                tax_exempt:
                  type: boolean
                effective_tax_rate_bps:
                  type: integer
                  nullable: true
                  minimum: 0
                  maximum: 100000
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopProductTaxRecord"

    delete:
      tags: [Admin]
      summary: Remove per-product tax override
      operationId: adminDeleteShopProductTax
      security:
        - bearerAuth: []
      parameters:
        - name: shop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ShopProductTaxRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        shop_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        tax_exempt:
          type: boolean
        effective_tax_rate_bps:
          type: integer
          nullable: true
...